backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.13
numpy>=1.26.0
scipy>=1.11.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
bcrypt>=4.1.0
scikit-learn>=1.3.0
dlib-bin>=19.24.0
face_recognition>=1.3.0
